feat(aft): Add `generate goldens` command
Removes bespoke generation script for goldens and integrates it into `aft`.
