+++ packages/aft/lib/src/commands/generate/generate_command.dart
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import 'package:aft/aft.dart';
+import 'package:aft/src/commands/generate/generate_goldens_command.dart';
 import 'package:aft/src/commands/generate/generate_sdk_command.dart';
 import 'package:aft/src/commands/generate/generate_workflows_command.dart';
 
@@ -10,6 +11,7 @@ class GenerateCommand extends AmplifyCommand {
   GenerateCommand() {
     addSubcommand(GenerateSdkCommand());
     addSubcommand(GenerateWorkflowsCommand());
+    addSubcommand(GenerateGoldensCommand());
   }
 
   @override
