+++ packages/aft/pubspec.yaml
@@ -15,6 +15,7 @@ dependencies:
   checked_yaml: ^2.0.0
   cli_util: ^0.3.5
   collection: ^1.16.0
+  file: ^7.0.0
   git: ^2.0.0
   glob: ^2.1.0
   graphs: ^2.1.0
